Print number of successes for dramatic tasks in SW checks
"SW check" is a simple Savage Worlds roll with or without additive
modifier(s), e.g., `s10+2`.
Here number of successes is not really a part of expression result
(because of special nature of such expression: not every Savage Worlds
roll is "simple", and not every +/- operator expression is a Savage
Worlds roll with modifier). So, just recognize SW check in the end.